Modified reinforce_agent to use torch.distributions.Categorical instead to calculate the distribution.
Created a config file for the reinforce agent, with hidden_size, discount_factor, and learning_rate.

Created a reinforce_sim file to run the simulation for 5 runs of a 1000 episodes each, saving the results to a json in the \results directory.

Added the A2.pdf assignment description, and added a .gitignore to ignore pycache and other irrelevant files